SMT: Avoid unnecessary allocation during translation
In writeBytes, there is no need to unpack bytestring into a list in order to fold it.
Also, we don't have to call `exprToSMT` to translate index and byte into
a bitvector. Thus, `wrap` can be pure.

Similarly, when translating literals, we can also directly translate the
value, without unnecessary conversion.