SMT: Do not parse term in get-val response
We do not need the term, so there is no reason to parse it.
The SMT parser we are using (`smt2-parser`) seems to be quite
inefficient, so we should limit the amount of work we require it to do.
We are always asking for a single value, and we expect it to be a
number. We can easily extract the value part of the response and parse
only that.