SMT: Fix Keccak constraints to use definitions
Previously, the axioms we pass to the solver ("keccak assumptions")
would use fully expressed buffers, even though we have applied CSE
before and substituted the buffers with GVars in the rest of SMT.

The problem was that we would collect keccak expressions from the
unsimplified Props.

The solution here is to collect the Keccak expressions from the
simplified Props. However, we have to take care not lose information
about concrete Keccak expressions, which are simplified in the first
step.
Therefore we change the flow slightly:
1. Collect all concrete Keccak expressions
2. Simplify the props, including concretizing keccaks
3. Introduce definitions for buffers and stores
4. Collect all keccak expressions from simplified props, substituted
   buffers and stores, and add concrete keccaks from step 1
5. Add necessary axioms for all keccak expressions from step 4

This ensures that the axioms added in step 5 use definitions introduced
in step 3.
The effect is that the SMT-LIB script is much smaller because we do not
repeat large buffer expressions in every axiom we add (as we are adding
quite a lot).