Update symbolic benchmark runner
After recent changes, we start to get some reasonable numbers in the
symbolic execution performance benchmarks.

On my laptop these are the current numbers (note that these are rough):
- erc20: ~3.5 seconds
- vat: ~5 minutes
- deposit: NA (still allocating too much memory during exploration)
- uniV2: ~2.5 minutes
- blockchain tests: ~26 seconds

These are using Bitwuzla as the SMT solver.
There is still a lot of room for improvement, but it's a good start at
the moment.