Expr: Fix collecting reads from abstract storage
Previously, we would lose some storage reads, because `foldProp` uses
`<>` to join the results and the implementation for Map discards the
value from the second map if both maps have entries with the same key.
That's not the behaviour we want, we need to take the set union.

The proposed solution is to use a wrapper type (`newtype`) for which we
implement our own operation `<>`.