Fix missing information about concrete Keccak computations
Previously, we would lose information about concrete Keccak computations
during simplification.
This information is then missing when we create constraints on the
abstract Keccak computations in the program.

The suggested solution is to collect concrete Keccak computations and
re-add them to the Props after simplification.

We also need to remove the call to simplification in `assertProps`.
The propositions are simplified before (in `checkSatWithProps`) and also
afterwards, in `assertPropsHelper`. We need to ensure that the
concrete Keccak computations survive until they are collected in
`assertPropsHelper`.