Fix incorrect prop simplification involving comparison
Currently, the expression `GEq (Var "v") (Lit 0x2)` is rewritten to
`IsZero (LT (Var "v") (Lit 0x2))`.

Then, in `simplifyProp`, this rewriting was triggered:
`go (PEq (Lit 1) (IsZero (LT a b))) = PLT a b`.

However, this rewriting is incorrect, the result should be `PGEq a b`,
or equivalently `PLEq b a`.
We prefer the latter because comparisons on prop level are rewritten in
terms of `PLT` and `PLEq` anyway.

Same reasoning can be applied for the other rule,
`go (PEq (Lit 1) (IsZero (LEq a b))) = PLEq a b`,
The result should be `PGT a b`, or equivalently, `PLT b a`.
Note however, that this rule seems to be redundant currently, as the
inner expression rewriting always rewrites `IsZero (LEq a b)` to `LT b a`.