EVM: Split symbolic and concrete sstore handling
Previously, there was a single piece of code handling both variants.
However, this solution unnecessary complicates the code and comes with
performance hits.

In symbolic case, we do not model gas, so all the code related to
computation and refunding of gas is not needed.
This code was trying to perform expression simplifications, which we now
avoid altogether.

In concrete case, we know we have to have concrete values everywhere, so
we can directly force expresseions to concrete values and avoid
unnecessary case splitting.