Test: Fix parsing of contracts in ethereum-tests
Due to the changes in RPC cache, parsing of contracts in blockchain
tests was failing.
The solution here is to use a dedicated type, `RPCContract` for
serializing and deserializing contracts.
To fix the blockchain tests, we bring back the JSON parsing code that
was used there before.

As a next step, we could refactor also how contracts are handled in
these tests.